Split tests per function
